requirements.txt
Django==4.2
djangorestframework>=3.12,<4
django-widget-tweaks>=1.4.12
